+++ packages/turbo/src/types.ts
@@ -75,7 +75,7 @@ export type StradaMessages = {
   [event: string]: StradaMessage;
 };
 
-// list of methods available for RNVisitableView module
+// List of methods available for RNVisitableView module
 export type DispatchCommandTypes =
   | 'injectJavaScript'
   | 'reload'
